reorganise file layout based on 2018 edition
