+++ Cargo.toml
@@ -13,15 +13,6 @@ repository = "https://github.com/untitaker/mates.rs"
 
 edition = "2018"
 
-[lib]
-name = "mates"
-path = "src/mates/lib.rs"
-
-[[bin]]
-name = "mates"
-test = false
-doc = false
-
 [dependencies]
 uuid = { version = "0.5", features = ["v4"] }
 atomicwrites = "0.1"
@@ -29,6 +20,7 @@ email = "0.0"
 clap = "2.31"
 vobject = "0.2"
 cursive = "0.14"
+anyhow = "1.0.26"
 
 [build-dependencies]
 clap = "2.31"
