+++ build.rs.bak
@@ -1,12 +1,10 @@
 extern crate clap;
+extern crate mates;
 
 use std::env;
 use std::fs;
 
 use clap::Shell;
-
-#[path = "src/mates/app.rs"]
-mod app;
 
 fn main() {
     let outdir = match env::var_os("OUT_DIR") {
@@ -15,7 +13,7 @@ fn main() {
     };
     fs::create_dir_all(&outdir).unwrap();
 
-    let mut app = app::app();
+    let mut app = mates::app();
     app.gen_completions("mates", Shell::Bash, &outdir);
     app.gen_completions("mates", Shell::Fish, &outdir);
     app.gen_completions("mates", Shell::Zsh, &outdir);
