--- src/bin/mates.rs
@@ -1,6 +1,0 @@
-extern crate mates;
-use mates::cli;
-
-fn main() {
-    cli::cli_main()
-}
